+++ packages/suite/src/actions/settings/deviceSettingsActions.ts
@@ -1,5 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import TrezorConnect from '@trezor/connect';
+import { analytics, EventType } from '@trezor/suite-analytics';
+
 import { addToast } from '@suite-actions/notificationActions';
 import * as suiteActions from '@suite-actions/suiteActions';
 import * as deviceUtils from '@suite-utils/device';
@@ -79,6 +81,9 @@ export const wipeDevice = () => async (dispatch: Dispatch, getState: GetState) =
             dispatch(suiteActions.forgetDevice(d));
         });
         dispatch(addToast({ type: 'device-wiped' }));
+        analytics.report({
+            type: EventType.SettingsDeviceWipe,
+        });
 
         // special case with webusb. device after wipe changes device_id. with webusb transport, device_id is used as path
         // and thus as descriptor for webusb. So, after device is wiped, in the transport layer, device is still paired